+++ src/economy/portdock.cc
@@ -350,11 +350,11 @@ void PortDock::log_general_info(Editor_Game_Base const & egbase)
 
 	Coords pos(m_warehouse->get_position());
 	molog
-		("PortDock for warehouse %u (at %i,%i) in fleet %u, need_ship: %s, waiting: %zu\n",
+		("PortDock for warehouse %u (at %i,%i) in fleet %u, need_ship: %s, waiting: %lu\n",
 		 m_warehouse ? m_warehouse->serial() : 0, pos.x, pos.y,
 		 m_fleet ? m_fleet->serial() : 0,
 		 m_need_ship ? "true" : "false",
-		 m_waiting.size());
+		 static_cast<long unsigned int>(m_waiting.size()));
 
 	container_iterate(std::vector<ShippingItem>, m_waiting, it) {
 		molog
